authorizer/requirements.txt
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.1
ecdsa==0.19.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
moto==5.0.17
packaging==24.1
pluggy==1.5.0
py-partiql-parser==0.5.5
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.4
xmltodict==0.13.0
